+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## unpublished
+### Fixed
+- Fixed `MetadataMatch.matrix()` raising `IndexError` in `equal_match` mode when no spectra match (e.g. comparing spectra from different instruments).
 
 ## [0.31.0] - 2025-10-06
 ### Added